requirements.txt
pandas>=1.5.0
jieba>=0.42.1
wordcloud>=1.9.2
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0  # 用于处理图片 